pyproject.toml
[project]
name = "linkedin-autopost"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "markdown>=3.9",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "youtube-transcript-api>=1.2.2",
]
